patch test2

+++ app/src/test/java/io/github/polysmee/database/databaselisteners/ListenerTest.java
@@ -14,6 +14,7 @@ import io.github.polysmee.database.databaselisteners.valuelisteners.StringValueL
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThrows;
+import static org.junit.Assert.assertTrue;
 
 public class ListenerTest {
 
@@ -21,7 +22,7 @@ public class ListenerTest {
     DownloadValueListener dv = (b) -> assertEquals(1, b.length);
     LoadValueListener lvl = (s) -> assertEquals("1234567890", s);
     StringValueListener sv = (s) -> assertEquals("1234567890", s);
-    LongValueListener lv = (l) -> assertEquals(1234567890, l);
+    LongValueListener lv = (l) -> assertEquals(1234567890L, (long) l);
     StringSetValueListener ssv = (s) -> assertEquals(0, s.size());
 
     @Test
@@ -30,7 +31,7 @@ public class ListenerTest {
         dv.onDone(new byte[]{0});
         lvl.onDone("1234567890");
         sv.onDone("1234567890");
-        lv.onDone(1234567890);
+        lv.onDone(1234567890L);
         ssv.onDone(new HashSet<>());
     }
